+++ src/others/files.c
@@ -9,7 +9,7 @@ int open_file(char *pathToResource, char *filename, FILE **f){
 	snprintf(fullPath, strlen(pathToResource)+strlen(filename)+1,"%s%s", pathToResource, filename);
 	if((*f=fopen(fullPath,"r"))==NULL){
 		free(fullPath);
-		return RETURN_ERROR;
+		return set_last_activity_error(OPENING_FILE_ERROR,"");
 	}
 	int entries=0;
 	char *line=NULL;
@@ -23,7 +23,7 @@ int open_file(char *pathToResource, char *filename, FILE **f){
 
 int open_file_str(char *pathToResource, char *filename, FILE **f, char ***s){
 	int entries=open_file(pathToResource, filename, f);
-	if(entries==RETURN_ERROR) return RETURN_ERROR;
+	if(entries==RETURN_ERROR) set_last_activity_error(OPENING_FILE_ERROR,"");
 	*s = (char**) malloc(entries * sizeof(char*)+1);
 	memset(*s,0,entries * sizeof(char*)+1);
 	size_t len=0;